+++ korge/src/jvmTest/kotlin/com/soywiz/korge/korge3d/main.kt
@@ -43,7 +43,7 @@ object Demo1 {
 				val angle = (tick / 4.0).degrees
 				camera.positionLookingAt(
 					cos(angle * 2) * 4, cos(angle * 3) * 4, -sin(angle) * 4, // Orbiting camera
-					0, 1, 0
+					0.0, 1.0, 0.0
 				)
 				tick++
 			}
@@ -102,7 +102,7 @@ object Demo2 {
 				val angle = (tick / 1.0).degrees
 				camera.positionLookingAt(
 					cos(angle * 1) * 4, 0.0, -sin(angle * 1) * 4, // Orbiting camera
-					0, 0, 0
+					0.0, 0.0, 0.0
 				)
 				tick++
 			}
@@ -115,7 +115,7 @@ object Demo2 {
 			val angle = 360.degrees * ratio
 			camera.positionLookingAt(
 				cos(angle) * distance, 0.0, sin(angle) * distance, // Orbiting camera
-				v.transform.translation.x, v.transform.translation.y, v.transform.translation.z
+				v.transform.translation.x.toDouble(), v.transform.translation.y.toDouble(), v.transform.translation.z.toDouble()
 			)
 		}
 	}
